refactor: remove legacy GitHub OAuth code
